webkit: Manage video playing state properly when browser goes to background
When the browser activity goes to background while video is playing in
fullscreen mode, two things happen around the same time. The native media
player resumes the video playback using the inline video view while the
BrowserActivity notifies WebView that activity is going to the background
which in turn pauses the video. Depending on the order of these calls,
the play call may happen after the pause which causes video to resume
while the browser activity is in the background.

This is fixed by handling the video state properly during exit
fullscreen so that the messages may arrive at any order and the browser
still behaves correctly.

Change-Id: Ie7e037ecde7663b8cac704e3ab56a064bd52f6cf
CRs-fixed: 451660